requirements.txt
openai
semantic-kernel
python-dotenv
requests
flask  